Move deleting session data from interrupt card to landing page, and updated related unit tests
