+++ src/controllers/landing.controller.ts
@@ -1,9 +1,17 @@
-import { Request, Response } from "express";
+import { NextFunction, Request, Response } from "express";
 
 import { logger } from "../utils/logger";
 import * as config from "../config";
+import { deleteApplicationData } from "../utils/application.data";
 
-export const get = (req: Request, res: Response) => {
-  logger.debugRequest(req, `GET LANDING_PAGE`);
-  return res.render(config.LANDING_PAGE);
+export const get = (req: Request, res: Response, next: NextFunction) => {
+  try {
+    logger.debugRequest(req, `GET LANDING_PAGE`);
+
+    deleteApplicationData(req.session);
+    return res.render(config.LANDING_PAGE);
+  } catch (error) {
+    logger.errorRequest(req, error);
+    next(error);
+  }
 };
